+++ tests/ModuleTest.php
@@ -95,16 +95,39 @@ HTML
         Yii::$app->set('cache', new FileCache(['cachePath' => '@yiiunit/runtime/cache']));
 
         $view = Yii::$app->view;
-        for($i=0;$i<=1;$i++){
+        for ($i = 0; $i <= 1; $i++) {
             ob_start();
             $module->logTarget->tag = 'tag' . $i;
-            if($view->beginCache(__FUNCTION__,['duration'=>3])){
+            if ($view->beginCache(__FUNCTION__, ['duration' => 3])) {
                 $module->renderToolbar(new Event(['sender' => $view]));
                 $view->endCache();
             }
             $output[$i] = ob_get_clean();
         }
-        $this->assertNotEquals($output[0],$output[1]);
+        $this->assertNotEquals($output[0], $output[1]);
+    }
+
+    /**
+     * Making sure debug toolbar does not error
+     * in case module ID is not "debug".
+     *
+     * @see https://github.com/yiisoft/yii2-debug/pull/176/
+     */
+    public function testToolbarWithCustomModuleID()
+    {
+        $moduleID = 'my_debug';
+
+        $module = new Module($moduleID);
+        $module->allowedIPs = ['*'];
+        Yii::$app->setModule($moduleID, $module);
+        $module->bootstrap(Yii::$app);
+        Yii::getLogger()->dispatcher = $this->getMock('yii\\log\\Dispatcher', ['dispatch']);
+
+        $view = Yii::$app->view;
+
+        ob_start();
+        $module->renderToolbar(new Event(['sender' => $view]));
+        ob_end_clean();
     }
 
     public function testDefaultVersion()