+++ views/default/panels/mail/detail.php
@@ -29,7 +29,7 @@ $listView->sorter = ['options' => ['class' => 'mail-sorter']];
 <div id="email-form" style="display: none;">
     <?php $form = ActiveForm::begin([
             'method' => 'get',
-            'action' => ['/debug/default/view', 'tag' => Yii::$app->request->get('tag'), 'panel' => 'mail'],
+            'action' => ['default/view', 'tag' => Yii::$app->request->get('tag'), 'panel' => 'mail'],
     ]); ?>
     <div class="row">
         <?= $form->field($searchModel, 'from', ['options' => ['class' => 'col-lg-6']])->textInput() ?>